Moved CurrentWeather.js to molecule folder location
Updated Apps.js import line
